Updated seeded system generation code
This code generates a crystal seeded system, with a seed close to the center of the box and random atomic arrangement outside the box. 
The main difference here is that the crystal does not have to be spherical. A regular replicated unit cell of the crystal can be taken as a seed and the atoms around could be thrown in random.